All encoding algorithms supported but the SAT-based. Code refactored.
