+++ encoding/includes_defines.h
@@ -1,0 +1,25 @@
+// libraries import
+#include <cstdio>
+#include <cstdlib>
+#include <cstring>
+#include <string>
+#include <vector>
+#include <math.h>
+#include <sstream>
+#include <map>
+
+// constant definition
+#define MAX_VERT	10000	//Max number of vertices of CPOG
+#define FILENAME_LENGTH	300	// Max length of file names
+#define MAX_LOG		10000	//2^(MAX_LOG) as max number of log_2 computable
+#define MAX_NAME	100	//Max length name for each node
+#define MAX_CPOG	1000	//Max number of graphs of CPOG
+#define MAX_WEIGHT	100000000//Used for finding min weight
+#define MAX_MEMORY	100000    //Max memory available for Workcraft
+
+#define stringLimit	1000
+#define eventsLimit	1000
+#define scenariosLimit	700
+#define predicatesLimit	200
+
+using namespace std;
